EM-6303 refactor service package
