+++ src/main/java/uk/gov/hmcts/reform/em/hrs/service/impl/ShareAndNotifyServiceImpl.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.em.hrs.service;
+package uk.gov.hmcts.reform.em.hrs.service.impl;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -14,6 +14,10 @@ import uk.gov.hmcts.reform.em.hrs.exception.ValidationErrorException;
 import uk.gov.hmcts.reform.em.hrs.model.CaseDocument;
 import uk.gov.hmcts.reform.em.hrs.model.CaseHearingRecording;
 import uk.gov.hmcts.reform.em.hrs.repository.HearingRecordingRepository;
+import uk.gov.hmcts.reform.em.hrs.service.Constants;
+import uk.gov.hmcts.reform.em.hrs.service.NotificationService;
+import uk.gov.hmcts.reform.em.hrs.service.ShareAndNotifyService;
+import uk.gov.hmcts.reform.em.hrs.service.ShareeService;
 import uk.gov.hmcts.reform.em.hrs.service.ccd.CaseDataContentCreator;
 import uk.gov.hmcts.reform.em.hrs.util.EmailValidator;
 import uk.gov.service.notify.NotificationClientException;
