+++ src/main/java/uk/gov/hmcts/reform/em/hrs/service/impl/TtlServiceImpl.java
@@ -1,8 +1,9 @@
-package uk.gov.hmcts.reform.em.hrs.service;
+package uk.gov.hmcts.reform.em.hrs.service.impl;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.em.hrs.config.TTLMapperConfig;
+import uk.gov.hmcts.reform.em.hrs.service.TtlService;
 
 import java.time.LocalDate;
 import java.time.Period;
