+++ src/test/java/uk/gov/hmcts/reform/em/hrs/controller/HearingRecordingControllerTest.java
@@ -20,11 +20,11 @@ import uk.gov.hmcts.reform.em.hrs.dto.HearingRecordingDto;
 import uk.gov.hmcts.reform.em.hrs.exception.SegmentDownloadException;
 import uk.gov.hmcts.reform.em.hrs.exception.UnauthorisedServiceException;
 import uk.gov.hmcts.reform.em.hrs.repository.HearingRecordingSegmentRepository;
-import uk.gov.hmcts.reform.em.hrs.service.AuditEntryService;
 import uk.gov.hmcts.reform.em.hrs.service.Constants;
 import uk.gov.hmcts.reform.em.hrs.service.HearingRecordingService;
 import uk.gov.hmcts.reform.em.hrs.service.SegmentDownloadService;
 import uk.gov.hmcts.reform.em.hrs.service.ShareAndNotifyService;
+import uk.gov.hmcts.reform.em.hrs.service.impl.AuditEntryService;
 
 import java.io.IOException;
 import java.time.Clock;
